+++ examples/servers-within-virtual-server-group/outputs.tf
@@ -23,7 +23,27 @@ output "this_slb_master_slave_servers" {
   value       = module.slb.this_slb_master_slave_servers
 }
 
+output "this_slb_master_slave_server_group_id" {
+  description = "The ID of master slave server group."
+  value       = module.slb.this_slb_master_slave_server_group_id
+}
+
+output "this_slb_master_slave_server_group_name" {
+  description = "The name of master slave server group"
+  value       = module.slb.this_slb_master_slave_server_group_name
+}
+
 output "this_slb_virtual_servers" {
   description = "List of slb virtual servers"
   value       = module.slb.this_slb_virtual_servers
+}
+
+output "this_slb_virtual_server_group_id" {
+  description = "The ID of virtual server group"
+  value       = module.slb.this_slb_virtual_server_group_id
+}
+
+output "this_slb_virtual_server_group_name" {
+  description = "The name of virtual server group"
+  value       = module.slb.this_slb_virtual_server_group_name
 }